+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Fix TT mate scores being stored without root/position correction
 * Hide transposition table implementation details in its module
+* datagen now saves data in viriformat
 
 ## [6.0]